+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Unreleased
 
+### Changed
+- Rename `package-links` manifest section to `package_links`.
+
 ## 0.7.1 - 2018-12-07
 ### Fixed
 - Fix an issue where path dependencies would not properly resolve transitively.